+++ gammapy/background/__init__.py
@@ -1,4 +1,3 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Background estimation and modeling methods."""
 from .phase import *
-from .reflected import *
