+++ gammapy/data/tests/test_obs_summary.py
@@ -4,13 +4,13 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from regions import CircleSkyRegion
-from gammapy.background import ReflectedRegionsBackgroundEstimator
 from gammapy.data import (
     DataStore,
     ObservationStats,
     ObservationSummary,
     ObservationTableSummary,
 )
+from gammapy.spectrum import ReflectedRegionsBackgroundEstimator
 from gammapy.utils.testing import mpl_plot_check, requires_data, requires_dependency
