+++ gammapy/spectrum/__init__.py
@@ -7,5 +7,6 @@ from .crab import *
 from .dataset import *
 from .extract import *
 from .flux_point import *
+from .reflected import *
 from .sensitivity import *
 from .utils import *
