+++ gammapy/spectrum/tests/test_extract.py
@@ -5,10 +5,13 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.coordinates import Angle, SkyCoord
 from regions import CircleSkyRegion
-from gammapy.background import ReflectedRegionsBackgroundEstimator
 from gammapy.data import DataStore, ObservationStats
 from gammapy.maps import WcsGeom, WcsNDMap
-from gammapy.spectrum import SpectrumDatasetOnOff, SpectrumExtraction
+from gammapy.spectrum import (
+    ReflectedRegionsBackgroundEstimator,
+    SpectrumDatasetOnOff,
+    SpectrumExtraction,
+)
 from gammapy.utils.testing import (
     assert_quantity_allclose,
     requires_data,
